+++ frontend/src/Impression.jsx
@@ -1,0 +1,11 @@
+import "./Impression.scss";
+
+function Impression({impression}) {
+  return(
+    <div className="impression">
+      {impression.quantity}
+    </div>
+  )
+}
+
+export default Impression;
