+++ frontend/src/Impression.scss
@@ -1,0 +1,5 @@
+.impression {
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+  padding-left: 0.5rem;
+}
